1.群组功能 2.群组成功的测试用例 冒烟

+++ src/services/group.service.ts
@@ -13,7 +13,7 @@ import { GroupMemberStatus } from '@/enums'
 import { UserService } from './user.service'
 import { randomUUID } from 'crypto'
 import { Request } from 'express'
-import { virtualCurrentUser } from '@/util/common.util'
+import commonUtil, { pageSkip, virtualCurrentUser } from '@/util/common.util'
 
 @Injectable()
 export class GroupService {
@@ -34,7 +34,7 @@ export class GroupService {
 
   async create (data: Prisma.GroupCreateInput): Promise<Group> {
     // 创建群组 + 创建群成员
-    const currentUser = await this.userService.findById(data.ownerId)
+    const currentUser = await this.userService.findByUId(data.ownerId)
     if (currentUser === null) {
       throw new HttpException('error', HttpStatus.BAD_REQUEST)
     }
@@ -60,8 +60,10 @@ export class GroupService {
 
   async getGroupMembers (param: GroupMemberReq): Promise<BasePageResp<GroupMembers>> {
     const data = await this.prisma.groupMembers.findMany({
-      where: { groupId: { equals: param.id } },
-      skip: param.skip(), // 计算需要跳过的数据量
+      where: {
+        groupId: { equals: param.id }
+      },
+      skip: commonUtil.pageSkip(param), // 计算需要跳过的数据量
       take: param.limit, // 指定每页取多少条数据
       orderBy: {
         createdAt: 'asc' // 按照创建时间降序排序，你可以根据需要修改排序字段和顺序
@@ -101,6 +103,7 @@ export class GroupService {
 
   // 邀请加入群聊
   async inviteMember (param: GroupInviteJoinReq): Promise<any> {
+    const currentId = virtualCurrentUser()
     const uIds = param.items.map(i => i.uid)
     const existMembers = await this.prisma.groupMembers.findMany({
       where: {
@@ -108,8 +111,10 @@ export class GroupService {
         uid: { in: uIds }
       }
     })
-    if (existMembers.length <= 0) {
-      const users: User[] = await this.userService.findByIds(uIds)
+    const existIds = existMembers.map(u => u.uid)
+    const saveIds = commonUtil.arrayDifference(uIds, existIds)
+    if (saveIds.length > 0) {
+      const users: User[] = await this.userService.findByUIds(saveIds)
       const members: Prisma.GroupMembersCreateInput[] = users.map(u => {
         const member: Prisma.GroupMembersCreateInput = {
           id: randomUUID(),
@@ -117,7 +122,7 @@ export class GroupService {
           uid: u.unionId,
           encPri: '',
           encKey: '',
-          inviteUid: 'todo',
+          inviteUid: currentId,
           role: 3,
           joinType: 1,
           myAlias: u.name,
@@ -132,7 +137,11 @@ export class GroupService {
 
   // 踢出群聊
   async memberKickOut (param: GroupKickOutReq): Promise<any> {
-    // todo 管理员鉴权
+    const currentId = virtualCurrentUser()
+    if (param.uids.includes(currentId)) {
+      throw new HttpException('不可踢出自己', HttpStatus.BAD_REQUEST)
+    }
+    await this.checkGroupRole(param.id, currentId, [1, 2])
     return await this.prisma.groupMembers.deleteMany({
       where: {
         groupId: { equals: param.id },
@@ -167,7 +176,7 @@ export class GroupService {
     // const group: Group = this.findOne(param.id)
     const result = await this.prisma.group.update({
       where: { id: param.id },
-      data: { name: param.name }
+      data: { name: param.name, updatedAt: new Date() }
     })
     if (result === null) {
       throw new HttpException('error', HttpStatus.INTERNAL_SERVER_ERROR)
@@ -182,7 +191,7 @@ export class GroupService {
     // const group: Group = this.findOne(param.id)
     const result = await this.prisma.group.update({
       where: { id: param.id },
-      data: { avatar: param.avatar }
+      data: { avatar: param.avatar, updatedAt: new Date() }
     })
     if (result === null) {
       throw new HttpException('error', HttpStatus.INTERNAL_SERVER_ERROR)
@@ -199,7 +208,8 @@ export class GroupService {
       where: { id: param.id },
       data: {
         notice: param.notice,
-        noticeMd5: param.noticeMd5
+        noticeMd5: param.noticeMd5,
+        updatedAt: new Date()
       }
     })
     if (result === null) {
@@ -217,7 +227,8 @@ export class GroupService {
       where: { id: param.id },
       data: {
         desc: param.desc,
-        descMd5: param.descMd5
+        descMd5: param.descMd5,
+        updatedAt: new Date()
       }
     })
     if (result === null) {
@@ -228,10 +239,11 @@ export class GroupService {
 
   // 修改我在群里面的昵称
   async changeAlias (param: GroupChangeAliasReq): Promise<any> {
+    const uId = virtualCurrentUser()
     return await this.prisma.groupMembers.updateMany({
       where: {
         groupId: param.id,
-        uid: param.id
+        uid: uId
       },
       data: { myAlias: param.alias }
     })
@@ -239,6 +251,11 @@ export class GroupService {
 
   // 退出群聊
   async quitGroup (param: BaseIdReq): Promise<any> {
+    const uId = virtualCurrentUser()
+    const groupRole = await this.groupRole(param.id, uId)
+    if (groupRole === 1) {
+      throw new HttpException('群主退群前请先进行转让或直接解散群聊', HttpStatus.BAD_REQUEST)
+    }
     return await this.prisma.groupMembers.deleteMany({
       where: {
         groupId: param.id,
@@ -250,6 +267,18 @@ export class GroupService {
   // 退出多个群聊
   async quitGroupBatch (param: BaseIdsArrayReq): Promise<any> {
     const uId = virtualCurrentUser()
+    const ownerGroups = await this.prisma.groupMembers.findMany({
+      where: {
+        uid: uId,
+        role: 1
+      }
+    })
+    if (ownerGroups.length > 0) {
+      const groups = await this.prisma.group.findMany({
+        where: { id: { in: ownerGroups.map(m => m.groupId) } }
+      })
+      throw new HttpException('群' + groups.join(',') + '是群主，请直接解散群聊', HttpStatus.BAD_REQUEST)
+    }
     return await this.prisma.groupMembers.deleteMany({
       where: {
         groupId: { in: param.ids },
@@ -261,6 +290,18 @@ export class GroupService {
   // 退出多个群聊
   async quitGroupAll (): Promise<any> {
     const uId = virtualCurrentUser()
+    const ownerGroups = await this.prisma.groupMembers.findMany({
+      where: {
+        uid: uId,
+        role: 1
+      }
+    })
+    if (ownerGroups.length > 0) {
+      const groups = await this.prisma.group.findMany({
+        where: { id: { in: ownerGroups.map(m => m.groupId) } }
+      })
+      throw new HttpException('群' + groups.join(',') + '是群主，请直接解散群聊', HttpStatus.BAD_REQUEST)
+    }
     return await this.prisma.groupMembers.deleteMany({
       where: {
         uid: uId
@@ -271,7 +312,18 @@ export class GroupService {
   // 解散群组
   async dismissGroup (param: BaseIdsArrayReq): Promise<any> {
     const uId = virtualCurrentUser()
-
+    const ownerGroups = await this.prisma.groupMembers.findMany({
+      where: {
+        uid: uId,
+        role: { not: 1 }
+      }
+    })
+    if (ownerGroups.length > 0) {
+      const groups = await this.prisma.group.findMany({
+        where: { id: { in: ownerGroups.map(m => m.groupId) } }
+      })
+      throw new HttpException('群' + groups.join(',') + '不是群主，没有权限', HttpStatus.BAD_REQUEST)
+    }
     const member = await this.prisma.groupMembers.findMany({
       where: {
         groupId: { in: param.ids },
@@ -334,7 +386,7 @@ export class GroupService {
       }
     })
     if (member === null) {
-      const user = await this.userService.findById(param.uid)
+      const user = await this.userService.findByUId(param.uid)
       if (user != null) {
         const input: Prisma.GroupMembersCreateInput = this.userService.user2GroupMemberInput(user, param.id)
         input.role = 2
@@ -355,6 +407,9 @@ export class GroupService {
   // 移除管理员(转为普通成员)
   async removeGroupManager (param: GroupApplyJoinReq): Promise<any> {
     const uId = virtualCurrentUser()
+    if (param.uids.includes(uId)) {
+      throw new HttpException('不可包含自己', HttpStatus.BAD_REQUEST)
+    }
     await this.checkGroupRole(param.id, uId, [1])
     return await this.prisma.groupMembers.updateMany({
       where: {
@@ -367,8 +422,40 @@ export class GroupService {
     })
   }
 
+  //  请求加入群组
+  async requireJoin (param: BaseIdReq): Promise<void> {
+    const currentId = virtualCurrentUser()
+    const groupMember = await this.groupMemberById(param.id, currentId)
+    if (groupMember !== null) {
+      if (groupMember.status > 0) {
+        throw new HttpException('已经加入群组', HttpStatus.BAD_REQUEST)
+      } else {
+        throw new HttpException('请等待管理审核', HttpStatus.BAD_REQUEST)
+      }
+    }
+
+    const currentUser = await this.userService.findByUId(currentId)
+    if (currentUser === null) {
+      throw new HttpException('error', HttpStatus.BAD_REQUEST)
+    }
+    const member: Prisma.GroupMembersCreateInput = {
+      id: randomUUID(),
+      groupId: param.id,
+      uid: currentId,
+      encPri: '',
+      encKey: '',
+      inviteUid: currentId,
+      role: 3,
+      joinType: 1,
+      myAlias: currentUser.name,
+      status: CommonEnum.OFF,
+      banType: CommonEnum.ON
+    }
+    await this.prisma.groupMembers.create({ data: member })
+  }
+
   // 待审核申请列表 只有群主和管理员有权限
-  async applyList (param: GroupApplyJoinReq): Promise<GroupInfoItem[]> {
+  async applyList (param: BaseIdsArrayReq): Promise<GroupInfoItem[]> {
     const uId = virtualCurrentUser()
     const managedGroups = await this.prisma.groupMembers.findMany({
       where: {
@@ -401,7 +488,7 @@ export class GroupService {
   }
 
   // 我的申请列表
-  async myPendingApplyList (param: GroupApplyJoinReq): Promise<MineGroupInfoItem[]> {
+  async myPendingApplyList (param: BaseIdsArrayReq): Promise<MineGroupInfoItem[]> {
     const uId = virtualCurrentUser()
     const pendingList = await this.prisma.groupMembers.findMany({
       where: {
@@ -456,23 +543,26 @@ export class GroupService {
 
   // 群组权限检查
   async checkGroupRole (groupId: string, uId: string, roles: number[]): Promise<any> {
-    if (!roles.includes(await this.groupRole(groupId, uId))) {
+    const mRole = await this.groupRole(groupId, uId)
+    if (!roles.includes(mRole)) {
       throw new HttpException('没有群组权限', HttpStatus.BAD_REQUEST)
     }
   }
 
-  async checkMemberLimit (): Promise<any> {}
-
-  async groupRole (groupId: string, uId: string): Promise<number> {
-    const member = await this.prisma.groupMembers.findMany({
+  async groupMemberById (groupId: string, uId: string): Promise<GroupMembers | null> {
+    return await this.prisma.groupMembers.findFirst({
       where: {
         groupId: { equals: groupId },
         uid: { equals: uId }
       }
     })
-    if (member.length <= 0) {
+  }
+
+  async groupRole (groupId: string, uId: string): Promise<number> {
+    const member = await this.groupMemberById(groupId, uId)
+    if (member === null) {
       throw new HttpException('不在群组内', HttpStatus.BAD_REQUEST)
     }
-    return member[0].role
+    return member.role
   }
 }
